refactor(agent): isolate native skill audit bootstrap
